backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-decouple==3.8
PyPDF2==3.0.1
pdfminer.six==20231228
python-dotenv==1.0.1
pydantic==2.9.2
groq==0.4.1
pydantic[email]==2.9.2

# python -m venv resume_env
# resume_env\Scripts\activate

# uvicorn app.main:app --reload --port 8000
